--- a/src/input/Cargo.toml
+++ b/src/input/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "pistoncore-input"
-version = "0.13.0"
+version = "0.14.0"
 authors = ["bvssvni <bvssvni@gmail.com>"]
 keywords = ["keyboard", "mouse", "input", "piston"]
 description = "A structure for user input"