--- a/src/input/src/after_render.rs
+++ b/src/input/src/after_render.rs
@@ -1,7 +1,5 @@
-use std::any::Any;
-
-use GenericEvent;
-use AFTER_RENDER;
+use Event;
+use Input;
 
 /// After render arguments.
 #[derive(Copy, Clone, PartialEq, Debug)]
@@ -20,6 +18,7 @@ pub trait AfterRenderEvent: Sized {
     }
 }
 
+/* TODO: Enable when specialization gets stable.
 impl<T: GenericEvent> AfterRenderEvent for T {
     fn from_after_render_args(args: &AfterRenderArgs, old_event: &T) -> Option<Self> {
         GenericEvent::from_args(AFTER_RENDER, args as &Any, old_event)
@@ -38,6 +37,34 @@ impl<T: GenericEvent> AfterRenderEvent for T {
                 panic!("Expected AfterRenderArgs")
             }
         })
+    }
+}
+*/
+
+impl AfterRenderEvent for Input {
+    fn from_after_render_args(_args: &AfterRenderArgs, _old_event: &Input) -> Option<Self> {
+        None
+    }
+
+    fn after_render<U, F>(&self, mut _f: F) -> Option<U>
+        where F: FnMut(&AfterRenderArgs) -> U
+    {
+        None
+    }
+}
+
+impl<I> AfterRenderEvent for Event<I> {
+    fn from_after_render_args(args: &AfterRenderArgs, _old_event: &Event<I>) -> Option<Self> {
+        Some(Event::AfterRender(*args))
+    }
+
+    fn after_render<U, F>(&self, mut f: F) -> Option<U>
+        where F: FnMut(&AfterRenderArgs) -> U
+    {
+        match *self {
+            Event::AfterRender(ref args) => Some(f(args)),
+            _ => None
+        }
     }
 }