--- a/src/input/src/focus.rs
+++ b/src/input/src/focus.rs
@@ -1,6 +1,6 @@
 use std::any::Any;
 
-use { GenericEvent, FOCUS };
+use { Event, GenericEvent, Input, FOCUS };
 
 /// When window gets or loses focus
 pub trait FocusEvent: Sized {
@@ -15,7 +15,46 @@ pub trait FocusEvent: Sized {
     }
 }
 
+/* TODO: Enable when specialization gets stable.
 impl<T: GenericEvent> FocusEvent for T {
+    fn from_focused(focused: bool, old_event: &Self) -> Option<Self> {
+        GenericEvent::from_args(FOCUS, &focused as &Any, old_event)
+    }
+
+    fn focus<U, F>(&self, mut f: F) -> Option<U>
+        where F: FnMut(bool) -> U
+    {
+        if self.event_id() != FOCUS {
+            return None;
+        }
+        self.with_args(|any| {
+            if let Some(&focused) = any.downcast_ref::<bool>() {
+                Some(f(focused))
+            } else {
+                panic!("Expected bool")
+            }
+        })
+    }
+}
+*/
+
+impl FocusEvent for Input {
+    fn from_focused(focused: bool, _old_event: &Self) -> Option<Self> {
+        Some(Input::Focus(focused))
+    }
+
+    fn focus<U, F>(&self, mut f: F) -> Option<U>
+        where F: FnMut(bool) -> U
+    {
+        match *self {
+            Input::Focus(focused) => Some(f(focused)),
+            _ => None
+        }
+    }
+}
+
+// TODO: Add impl for `Event<Input>` when specialization gets stable.
+impl<I: GenericEvent> FocusEvent for Event<I> {
     fn from_focused(focused: bool, old_event: &Self) -> Option<Self> {
         GenericEvent::from_args(FOCUS, &focused as &Any, old_event)
     }