--- a/src/input/src/release.rs
+++ b/src/input/src/release.rs
@@ -1,6 +1,4 @@
-use std::any::Any;
-
-use { Button, Event, GenericEvent, Input, RELEASE };
+use { Button, Event, Input };
 
 /// The release of a button
 pub trait ReleaseEvent: Sized {
@@ -53,25 +51,23 @@ impl ReleaseEvent for Input {
     }
 }
 
-// TODO: Add impl for `Event<Input>` when specialization gets stable.
-impl<I: GenericEvent> ReleaseEvent for Event<I> {
+impl<I: ReleaseEvent> ReleaseEvent for Event<I> {
     fn from_button(button: Button, old_event: &Self) -> Option<Self> {
-        GenericEvent::from_args(RELEASE, &button as &Any, old_event)
+        if let &Event::Input(ref old_input) = old_event {
+            <I as ReleaseEvent>::from_button(button, old_input)
+                .map(|x| Event::Input(x))
+        } else {
+            None
+        }
     }
 
-    fn release<U, F>(&self, mut f: F) -> Option<U>
+    fn release<U, F>(&self, f: F) -> Option<U>
         where F: FnMut(Button) -> U
     {
-        if self.event_id() != RELEASE {
-            return None;
+        match *self {
+            Event::Input(ref x) => x.release(f),
+            _ => None
         }
-        self.with_args(|any| {
-            if let Some(&button) = any.downcast_ref::<Button>() {
-                Some(f(button))
-            } else {
-                panic!("Expected Button")
-            }
-        })
     }
 }