Closes https://github.com/PistonDevelopers/piston/issues/1105

Benchmarks shows significant improvement in performance.

```
Before:
test bench_event_after_render ... bench:           0 ns/iter (+/- 0)
test bench_event_focus ... bench:                 38 ns/iter (+/- 17)
test bench_input_focus ... bench:                 27 ns/iter (+/- 16)
test bench_event_idle ... bench:                   0 ns/iter (+/- 0)
test bench_event_mouse_cursor   ... bench:        58 ns/iter (+/- 19)
test bench_event_mouse_relative ... bench:        58 ns/iter (+/- 12)
test bench_event_mouse_scroll   ... bench:        67 ns/iter (+/- 37)
test bench_input_mouse_cursor   ... bench:        29 ns/iter (+/- 6)
test bench_input_mouse_relative ... bench:        34 ns/iter (+/- 7)
test bench_input_mouse_scroll   ... bench:        41 ns/iter (+/- 18)
test bench_event_press ... bench:                 40 ns/iter (+/- 8)
test bench_input_press ... bench:                 31 ns/iter (+/- 7)
test bench_event_release ... bench:               43 ns/iter (+/- 14)
test bench_input_release ... bench:               32 ns/iter (+/- 5)
test bench_event_render ... bench:                 0 ns/iter (+/- 0)
test bench_event_resize ... bench:                43 ns/iter (+/- 17)
test bench_input_resize ... bench:                35 ns/iter (+/- 9)
test bench_event_text ... bench:                 109 ns/iter (+/- 73)
test bench_input_text ... bench:                  92 ns/iter (+/- 33)
test bench_event_update ... bench:                 0 ns/iter (+/- 0)

After:

test bench_event_after_render ... bench:           0 ns/iter (+/- 0)
test bench_event_focus ... bench:                  2 ns/iter (+/- 0)
test bench_input_focus ... bench:                  2 ns/iter (+/- 0)
test bench_event_idle ... bench:                   0 ns/iter (+/- 0)
test bench_event_mouse_cursor   ... bench:         2 ns/iter (+/- 1)
test bench_event_mouse_relative ... bench:         2 ns/iter (+/- 1)
test bench_event_mouse_scroll   ... bench:         2 ns/iter (+/- 0)
test bench_input_mouse_cursor   ... bench:         2 ns/iter (+/- 0)
test bench_input_mouse_relative ... bench:         2 ns/iter (+/- 0)
test bench_input_mouse_scroll   ... bench:         2 ns/iter (+/- 0)
test bench_event_press ... bench:                  3 ns/iter (+/- 0)
test bench_input_press ... bench:                  3 ns/iter (+/- 0)
test bench_event_release ... bench:                3 ns/iter (+/- 0)
test bench_input_release ... bench:                3 ns/iter (+/- 1)
test bench_event_render ... bench:                 0 ns/iter (+/- 0)
test bench_event_resize ... bench:                 2 ns/iter (+/- 0)
test bench_input_resize ... bench:                 2 ns/iter (+/- 0)
test bench_event_text ... bench:                  31 ns/iter (+/- 1)
test bench_input_text ... bench:                  31 ns/iter (+/- 7)
test bench_event_update ... bench:                 0 ns/iter (+/- 0)
```